refactor: Repository로 네이밍 변경 시 누락된 테스트 파일명·패키지 수정
