+++ src/test/java/roomescape/repository/ReservationQueryingRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.repository;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -16,7 +16,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @JdbcTest
 @Import(ReservationQueryingRepository.class)
-public class QueryingDaoTest {
+public class ReservationQueryingRepositoryTest {
 
     @Autowired
     private ReservationQueryingRepository reservationQueryingRepository;
